Switch from 's' element to 'strike'
I don't fully understand what is going on but this fixes strikethrough on Chrome + bulbs-cms. Since scribe is deprecated
not going to spend any more time tracking this down.

It seems as though something inside the editor <div> is automatically deleting <s> tags. This can be seen by manually changing an <s>
tag to a <strike> tag within dev tools (though it doesn't ALWAYS seem to do it in the editor project, which is why this
bug is so maddening). However it does seem to always change <s> to <strike> in bulbs-cms and the actual CMS apps.